# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Techno Support VMS API
  version: 1.0.0
  description: |
    REST configuration and control API for Techno Support Video Management System.
    Base URL: /api/v1
    
    Authentication: Bearer Token (JWT)
paths:
...
  /cameras:
    get:
      summary: List cameras
      tags: [Cameras]
      parameters:
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
        - name: page_token
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of cameras
          content:
            application/json:
              schema:
                type: object
                properties:
                  cameras:
                    type: array
                    items: { $ref: '#/components/schemas/Camera' }
                  next_page_token: { type: string }
    post:
      summary: Create camera
      tags: [Cameras]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCameraRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Camera' }
...
components:
  schemas:
    Camera:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        ip: { type: string }
        status: { type: string, enum: [ONLINE, OFFLINE] }
    CreateCameraRequest:
      type: object
      required: [name, ip, username, password]
      properties:
        name: { type: string }
        ip: { type: string }
        username: { type: string }
        password: { type: string }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []